budget_alto_os/requirements.txt
aiohttp==3.8.5
azure-iot-device==2.6.0
azure-iot-hub==2.6.1
BAC0==22.9.21
firebase==4.0.1
flask==2.3.2
gcloud==0.18.3
getmac==0.9.4
Jinja2==3.1.2
numpy==1.26.4
pandas==2.0.3
pendulum==2.1.2
psycopg2-binary==2.9.7
pycrypto==2.6.1
PyJWT==2.8.0
pymodbus==3.3.2
pymongo==4.6.2
Pyrebase4==4.6.0
pyserial==3.5
python-dotenv==1.0.1
python_jwt==4.0.0
requests==2.23.0
requests_toolbelt==0.10.1
sseclient==0.0.27
supabase==2.6.0
ws4py==0.5.1
PyYAML
wheel==0.30
asyncio==3.4.3
pythermalcomfort==3.3.0